додала врапер і валідацію помилки
