+++ controllers/contactsControllers/getContactById.js
@@ -2,18 +2,14 @@ const contacts = require("../../models/contacts");
 const { HttpError } = require("../../helpers");
 
 const getContactById = async (req, res, next) => {
-  try { 
+   
     const { id } = req.params;
     const result = await contacts.getContactById(id);
     if (!result) {
       throw HttpError(404, "Not found");
     }
     res.json(result)
-  }
-  catch (error) {
-    next(error);
-  
-  }
+ 
 }
 
 module.exports = getContactById
