+++ controllers/contactsControllers/listContacts.js
@@ -1,13 +1,10 @@
 const contacts = require("../../models/contacts")
 const listContacts = async (req, res, next) => {
-  try {
+  
     const result = await contacts.listContacts();
 
   res.json(result)
-  }
-  catch (error) {
-   next(error)
-  }
+ 
 }
 
 module.exports = listContacts
