+++ controllers/contactsControllers/update.js
@@ -3,21 +3,18 @@ const { HttpError } = require("../../helpers");
 const addShema = require("../../schemas/contacts");
 
  const updateContact = async (req, res, next) => {
-  try {
-   const { error } = addShema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message)
-    }
+  
+  //  const { error } = addShema.validate(req.body);
+  //   if (error) {
+  //     throw HttpError(400, error.message)
+  //   }
     const { id } = req.params;
     const result = await contacts.updateContact(id, req.body);
     if (!result) {
       throw HttpError(404, "Not founded")
     }
     res.json(result)
-  }
-  catch (error) {
-    next(error)
-  }
+
  }
 
  module.exports = updateContact
