+++ helpers/ctrlWrappers.js
@@ -1,7 +1,7 @@
 const ctrlWrappers = ctrl => {
     const func = async (req, res, next) => {
         try {
-           await (req, res, next) 
+           await ctrl(req, res, next) 
         }
         catch (error) {
             next(error)
