+++ middlewares/index.js
@@ -1,0 +1,5 @@
+const validateBody = require("./validateBody");
+
+module.exports = {
+    validateBody
+}
